# api/v2/swagger/swagger.yaml
definitions:
  api.BusinessError:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
  api.Response:
    properties:
      data:
        type: object
      response:
        $ref: '#/definitions/api.BusinessError'
    type: object
...
  jsoniter.Any:
    type: object
  translation.BundleData:
    properties:
      component:
        type: string
      locale:
        type: string
      messages:
        $ref: '#/definitions/jsoniter.Any'
    required:
    - component
    - locale
    - messages
    type: object
  translation.UpdateBundle:
    properties:
      creation:
        properties:
          operationid:
            type: string
        type: object
      dataOrigin:
        type: string
      productName:
        type: string
      translation:
        items:
          $ref: '#/definitions/translation.BundleData'
        type: array
      version:
        type: string
    required:
    - productName
    - translation
    - version
    type: object
  translation.UpdateTranslationDTO:
    properties:
      data:
        $ref: '#/definitions/translation.UpdateBundle'
      requester:
        type: string
    required:
    - data
    type: object
info:
  contact:
    name: Singleton Team
    url: https://github.com/vmware/singleton
  description: This is a Singleton server
  license:
    name: EPL-2.0
    url: https://www.eclipse.org/legal/epl-2.0/
  title: Singleton Service API
  version: "2.0"
paths:
...
  /translation/products/{productName}/versions/{version}:
    get:
      description: Get messages of all the product or parts of the product
      parameters:
      - description: product name
        in: path
        name: productName
        required: true
        type: string
      - description: version
        in: path
        name: version
        required: true
        type: string
      - description: locales
        in: query
        name: locales
        type: string
      - description: components
        in: query
        name: components
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response'
        "206":
          description: Successful Partially
          schema:
            $ref: '#/definitions/api.Response'
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get messages of the product
      tags:
      - translation-product-api
    put:
      consumes:
      - application/json
      description: Update bundles
      parameters:
      - description: product name
        in: path
        name: productName
        required: true
        type: string
      - description: version
        in: path
        name: version
        required: true
        type: string
      - description: translationData
        in: body
        name: translationData
        required: true
        schema:
          $ref: '#/definitions/translation.UpdateTranslationDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response'
        "206":
          description: Successful Partially
          schema:
            $ref: '#/definitions/api.Response'
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Update bundles
      tags:
      - translation-sync-api
  /translation/products/{productName}/versions/{version}/componentlist:
    get:
      description: Get available component names in the product
      parameters:
      - description: product name
        in: path
        name: productName
        required: true
        type: string
      - description: version
        in: path
        name: version
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response'
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get component names
      tags:
      - translation-product-api
  /translation/products/{productName}/versions/{version}/localelist:
    get:
      description: Get available locale names in the product
      parameters:
      - description: product name
        in: path
        name: productName
        required: true
        type: string
      - description: version
        in: path
        name: version
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response'
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get locale names
      tags:
      - translation-product-api
  /translation/products/{productName}/versions/{version}/locales/{locale}/components/{component}:
    get:
      description: Get messages of a single bundle
      parameters:
      - description: product name
        in: path
        name: productName
        required: true
        type: string
      - description: version
        in: path
        name: version
        required: true
        type: string
      - description: locale name
        in: path
        name: locale
        required: true
        type: string
      - description: component name
        in: path
        name: component
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response'
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get messages of a single bundle
      tags:
      - translation-product-component-api
  /translation/products/{productName}/versions/{version}/locales/{locale}/components/{component}/keys/{key}:
    get:
      deprecated: true
      description: Get a message by its key
      parameters:
      - description: product name
        in: path
        name: productName
        required: true
        type: string
      - description: version
        in: path
        name: version
        required: true
        type: string
      - description: locale name
        in: path
        name: locale
        required: true
        type: string
      - description: component name
        in: path
        name: component
        required: true
        type: string
      - description: key
        in: path
        name: key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response'
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get a message
      tags:
      - translation-product-component-key-api
swagger: "2.0"
...